telemetry reader

+++ src/openvslam/gps/data.cc
@@ -14,7 +14,7 @@ data::data(const double latitude, const double longitude, const double height,
 }
 
 data::data(const Vec3_t& llh,  const double dop_precision,
-           const int fix, const int speed_2d, const int speed_3d,
+           const int fix, const double speed_2d, const double speed_3d,
            const double ts)
     : llh_(llh), dop_precision_(dop_precision),
       speed_2d_(speed_2d), speed_3d_(speed_3d), fix_(fix), ts_(ts) {
